pm-interview-coach/scripts/requirements.txt
pytubefix>=6.0.0
openai-whisper>=20230918
openai>=1.12.0
sounddevice>=0.4.6
scipy>=1.11.0
flask>=3.0.0
